#17 [refactor] : mvvm 세팅
