--- app/src/main/java/com/sopt/withsuhyeon/core/base/UiEvent.kt
@@ -1,3 +1,0 @@
-package com.sopt.withsuhyeon.core.base
-
-interface UiEvent
