--- app/src/main/java/com/sopt/withsuhyeon/core/base/UiState.kt
@@ -1,3 +1,0 @@
-package com.sopt.withsuhyeon.core.base
-
-interface UiState
