+++ app/src/main/java/com/sopt/withsuhyeon/feature/chat/ChatViewModel.kt
@@ -1,23 +1,11 @@
 package com.sopt.withsuhyeon.feature.chat
 
-import com.sopt.withsuhyeon.core.base.BaseViewModel
-import com.sopt.withsuhyeon.feature.chat.contract.ChatContract.ChatEvent
-import com.sopt.withsuhyeon.feature.chat.contract.ChatContract.ChatSideEffect
-import com.sopt.withsuhyeon.feature.chat.contract.ChatContract.ChatUiState
+import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class ChatViewModel @Inject constructor(
-) : BaseViewModel<ChatUiState, ChatSideEffect, ChatEvent>() {
-    override fun createInitialState(): ChatUiState {
-        return ChatUiState()
-    }
-
-    override suspend fun handleEvent(event: ChatEvent) {
-        when(event) {
-            else -> {}
-        }
-    }
+) : ViewModel() {
 
 }
