+++ app/src/main/java/com/sopt/withsuhyeon/feature/dummy/base/UiEvent.kt
@@ -1,0 +1,3 @@
+package com.sopt.withsuhyeon.feature.dummy.base
+
+interface UiEvent
