+++ app/src/main/java/com/sopt/withsuhyeon/feature/home/HomeViewModel.kt
@@ -1,23 +1,11 @@
 package com.sopt.withsuhyeon.feature.home
 
-import com.sopt.withsuhyeon.core.base.BaseViewModel
+import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
-import com.sopt.withsuhyeon.feature.home.contract.HomeContract.HomeUiState
-import com.sopt.withsuhyeon.feature.home.contract.HomeContract.HomeSideEffect
-import com.sopt.withsuhyeon.feature.home.contract.HomeContract.HomeEvent
 
 @HiltViewModel
 class HomeViewModel @Inject constructor(
-) : BaseViewModel<HomeUiState, HomeSideEffect, HomeEvent>() {
+) : ViewModel() {
 
-    override fun createInitialState(): HomeUiState {
-        return HomeUiState()
-    }
-
-    override suspend fun handleEvent(event: HomeEvent) {
-        when (event) {
-            else -> {}
-        }
-    }
 }
