--- app/src/main/java/com/sopt/withsuhyeon/feature/onboarding/contract/OnBoardingContract.kt
@@ -1,19 +1,0 @@
-package com.sopt.withsuhyeon.feature.onboarding.contract
-
-import com.sopt.withsuhyeon.core.base.UiEvent
-import com.sopt.withsuhyeon.core.base.UiSideEffect
-import com.sopt.withsuhyeon.core.base.UiState
-
-class OnBoardingContract {
-    data class OnBoardingUiState(
-        val isLoading: Boolean = false,
-    ) : UiState
-
-    sealed interface OnBoardingSideEffect : UiSideEffect {
-        data object NavigateToHome : OnBoardingSideEffect
-    }
-
-    sealed interface OnBoardingEvent : UiEvent {
-        data object EndLoading : OnBoardingEvent
-    }
-}
